@@ -63,10 +63,10 @@
     "istanbul": "^0.4",
     "jsdoc": "^3.0",
     "memory-fs": "^0.4.1",
-    "mocha": "^5.0.5",
+    "mocha": "^5.1.1",
     "rimraf": "^2.0",
     "superagent": "^3.8.2",
-    "webpack": "^4.4.1",
-    "webpack-dev-server": "^3.1.1"
+    "webpack": "^4.6.0",
+    "webpack-dev-server": "^3.1.3"
   }
 }
